Updated Material Design
Updated the material design in general to match the specs
set for this project by Udacity. App uses a coordinator
layout and appBarLayout to handle the toolbar and scrolling
content. The font is consistent and is the default Android
font. Colors are consistent, coherent, and make viewing and
using the app easy. Elevations were set according to
recommendations in Google's material design docs. Spacing
and alignment for various views looks much cleaner.